requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
pydantic==2.5.2

# Streamlit Dependencies
streamlit==1.31.1
Pillow==10.2.0
opencv-python==4.8.1.78
numpy==1.26.2
fpdf2==2.7.6
torch==2.1.1
torchvision==0.16.1

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Development Dependencies
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Additional Dependencies
tqdm==4.66.1
aiofiles==23.2.1
